refactor: move .js files to control folder
