+++ src/control/drawing.ts
